+++ module-system/module-implementations/sov-evm/src/tests/mod.rs
@@ -1,2 +1,3 @@
 mod call_tests;
+mod cfg_tests;
 mod tx_tests;
